YARN-10548. Decouple AM runner logic from SLSRunner. Contributed by Szilard Nemeth.
